# docs/swagger.json
{
	"components": {
...
		"schemas": {
			"HTTPError": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double"
					},
					"details": {
						"type": "string"
					},
					"additional_info": {}
				},
				"required": [
					"code",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Result": {
				"properties": {
					"result": {},
					"error": {
						"$ref": "#/components/schemas/HTTPError"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"errorMsg": {
				"properties": {
					"message": {
						"type": "string"
					},
					"additional_info": {}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ITopic": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"imageId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IRefreshToken": {
				"properties": {
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IEmail": {
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IResetPw": {
				"properties": {
					"password1": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"password1",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"accessToken": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/locations": {
			"get": {
				"operationId": "GetLocations",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								},
								"examples": {
									"Example 1": {
										"value": {
											"data": [
												{
													"id": 1,
													"city": "Bandung",
													"stateProvince": "Jawa Barat",
													"country": "Indonesia",
													"postalCode": "40135",
													"createdAt": "2022-04-14T08:06:40.294Z",
													"updatedAt": "2022-04-14T08:06:40.294Z"
												},
												{
													"id": 2,
													"city": "Jakarta",
													"stateProvince": "DKI Jakarta",
													"country": "Indonesia",
													"postalCode": "40136",
													"createdAt": "2022-04-14T08:06:40.294Z",
													"updatedAt": "2022-04-14T08:06:40.294Z"
												}
											]
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Could not get locations, ask administrator"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get all locations",
				"tags": [
					"Location"
				],
				"security": [],
				"parameters": []
			}
		},
		"/topics": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								},
								"examples": {
									"Example 1": {
										"value": {
											"data": {
												"name": "Topic 1"
											}
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Server error, cannot create topic"
										}
									}
								}
							}
						}
					}
				},
				"description": "render",
				"tags": [
					"Topic"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ITopic"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetTopics",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								},
								"examples": {
									"Example 1": {
										"value": {
											"data": [
												{
													"id": 1,
													"name": "programming",
													"description": "create a program",
													"createdAt": "2022-04-08T06:13:51.594Z",
													"updatedAt": "2022-04-08T06:13:51.594Z",
													"imageId": null
												},
												{
													"id": 2,
													"name": "math",
													"description": "calculate",
													"createdAt": "2022-04-08T06:13:51.594Z",
													"updatedAt": "2022-04-08T06:13:51.594Z",
													"imageId": null
												}
											]
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Topic not found"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get all topics",
				"tags": [
					"Topic"
				],
				"security": [],
				"parameters": []
			}
		},
...
		"/users/refresh": {
			"post": {
				"operationId": "RefreshToken",
				"responses": {
					"200": {
						"description": "Refresh token success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								},
								"examples": {
									"Example 1": {
										"value": {
											"accessToken": "secrettoken",
											"refreshToken": "secrettoken"
										}
									}
								}
							}
						}
					},
					"403": {
						"description": "Invalid token",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Invalid refresh token"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Input non valid",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Failed during input validation"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Could not refresh"
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IRefreshToken"
							}
						}
					}
				}
			}
		},
...
		"/users/forget-password": {
			"post": {
				"operationId": "ForgetPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "User not found"
										}
									}
								}
							}
						}
					}
				},
				"description": "Send password reset to user's email",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IEmail"
							}
						}
					}
				}
			}
		},
		"/users/reset-password/{UID}/{token}": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								}
							}
						}
					},
					"400": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Passwords don't match"
										}
									}
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Invalid token"
										}
									}
								}
							}
						}
					}
				},
				"description": "Reset user's password",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"description": "the user's ID",
						"in": "path",
						"name": "UID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "the user's token",
						"in": "path",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IResetPw"
							}
						}
					}
				}
			}
		},
		"/users/profile": {
			"get": {
				"operationId": "GetProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result",
									"properties": {
										"id": {
											"type": "number"
										},
										"username": {
											"type": "string"
										},
										"fullname": {
											"type": "string"
										},
										"email": {
											"type": "string"
										},
										"lastLogin": {
											"type": "string",
											"format": "datetime"
										},
										"hasImage": {
											"type": "object",
											"properties": {
												"fileName": {
													"type": "string"
												},
												"mimeType": {
													"type": "string"
												},
												"url": {
													"type": "string"
												},
												"size": {
													"type": "number"
												},
												"extension": {
													"type": "string"
												}
											}
										},
										"locatedOn": {
											"type": "object",
											"properties": {
												"city": {
													"type": "string"
												},
												"stateProvince": {
													"type": "string"
												},
												"country": {
													"type": "string"
												},
												"postalCode": {
													"type": "number"
												}
											}
										},
										"interestedIn": {
											"type": "array",
											"items": {
												"type": "object",
												"properties": {
													"name": {
														"type": "string"
													},
													"description": {
														"type": "string"
													},
													"imageId": {
														"type": "number"
													}
												}
											}
										}
									}
								},
								"examples": {
									"Example 1": {
										"value": {
											"id": 7,
											"fullname": "Ini Nama Panjang",
											"username": "username",
											"email": "semangatintegrasi@gmail.com",
											"lastLogin": "2022-04-01T18:46:19.576Z",
											"hasImage": null,
											"locatedOn": {
												"city": "Bandung",
												"stateProvince": "Washington AC",
												"country": "Inazuma",
												"postalCode": "81451",
												"createdAt": "2022-04-01T18:37:13.474Z",
												"updatedAt": "2022-04-01T18:37:13.474Z",
												"imageId": null
											},
											"interestedIn": {
												"name": "Genshin Tidak Impact",
												"description": "Yes",
												"createdAt": "2022-04-01T18:37:13.474Z",
												"updatedAt": "2022-04-01T18:37:13.474Z",
												"imageId": null
											},
											"joinedIn": "2022-04-01T18:37:13.474Z"
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "User not authorized"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Could not get profile, ask administrator"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get user's profile detail",
				"tags": [
					"User"
				],
				"security": [
					{
						"accessToken": []
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}